+++ src/System.cpp
@@ -7,7 +7,7 @@ namespace ORB_VISLAM
 {
 
 System::System(	const string &settingFilePath, float scale,
-						int win_sz, float ssd_th, float ssd_ratio_th, int minFeat)
+						int win_sz, float ssd_th, float ssd_ratio_th, size_t minFeat)
 {
 	cout << "" << endl;
 	cout << "#########################################################################" << endl;
